requirements.txt
anyio==4.4.0
boto3==1.34.122
boto3-stubs==1.34.123
botocore==1.34.122
botocore-stubs==1.34.123
certifi==2024.6.2
charset-normalizer==3.3.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jmespath==1.0.1
mypy-boto3-cloudformation==1.34.111
mypy-boto3-dynamodb==1.34.114
mypy-boto3-ec2==1.34.119
mypy-boto3-lambda==1.34.77
mypy-boto3-rds==1.34.116
mypy-boto3-s3==1.34.120
mypy-boto3-sqs==1.34.121
python-dateutil==2.9.0.post0
python-telegram-bot==21.2
requests==2.32.3
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
types-awscrt==0.20.11
types-s3transfer==0.10.1
urllib3==2.2.1
wcwidth==0.2.13
